supplier service

+++ LPGManager/Interfaces/SupplierInterface/ISupplierService.cs
@@ -5,10 +5,10 @@ namespace LPGManager.Interfaces.SupplierInterface
     public interface ISupplierService
     {
 
-        Task<IEnumerable<Supplier>> GetAllAsync();        
-        Task<Supplier> GetAsync(int id);
+        Task<IEnumerable<Supplier>> GetAllAsync();
+        Task<Supplier> GetAsync(long id);
         Task<Supplier> AddAsync(Supplier supplier);
         Task<Supplier> UpdateAsync(Supplier model);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
